+++ lib/modules/tab_5/views/output_screen.dart
@@ -57,14 +57,44 @@ class Tab5OutputScreen extends ConsumerWidget {
                       shrinkWrap: true,
                       itemBuilder: (context, index) {
                         return Dismissible(
+                          // https://stackoverflow.com/questions/64135284/how-to-achieve-delete-and-undo-operations-on-dismissible-widget-in-flutter
+                          confirmDismiss: (direction) async {
+                            if (direction == DismissDirection.startToEnd) {
+                              return await showDialog(
+                                    context: context,
+                                    builder: (context) {
+                                      return AlertDialog(
+                                        title: const Text("Delete"),
+                                        content: const Text(
+                                            'Are you sure you want to delete?'),
+                                        actions: [
+                                          IconButton.filledTonal(
+                                              icon: const Icon(Icons.done),
+                                              onPressed: () =>
+                                                  Navigator.pop(context, true)),
+                                          IconButton.filled(
+                                              icon: const Icon(Icons.dangerous),
+                                              onPressed: () => Navigator.pop(
+                                                  context, false)),
+                                        ],
+                                      );
+                                    },
+                                  ) ??
+                                  false;
+                            } else {
+                              return false;
+                            }
+                          },
                           key: Key(data[index].date),
                           background: Container(color: Colors.red),
                           onDismissed: (direction) {
-                            ref
-                                .read(tab5RepositoryProvider.notifier)
-                                .deleteModel(data[index]);
-                            data.removeAt(index);
-                            setState(() {});
+                            if (direction == DismissDirection.startToEnd) {
+                              ref
+                                  .read(tab5RepositoryProvider.notifier)
+                                  .deleteModel(data[index]);
+                              data.removeAt(index);
+                              setState(() {});
+                            }
                           },
                           child: InkWell(
                             onTap: () {
